Extract the C++ function calling part in `overload_resolution` to `call`.
