+++ toolchain/check/BUILD
@@ -21,6 +21,7 @@ cc_library(
         "context.cpp",
         "control_flow.cpp",
         "convert.cpp",
+        "cpp/call.cpp",
         "cpp/custom_type_mapping.cpp",
         "cpp/import.cpp",
         "cpp/location.cpp",
@@ -72,6 +73,7 @@ cc_library(
         "context.h",
         "control_flow.h",
         "convert.h",
+        "cpp/call.h",
         "cpp/custom_type_mapping.h",
         "cpp/import.h",
         "cpp/location.h",
