+++ toolchain/check/cpp/overload_resolution.cpp
@@ -11,6 +11,7 @@
 #include "toolchain/check/call.h"
 #include "toolchain/check/cpp/import.h"
 #include "toolchain/check/cpp/location.h"
+#include "toolchain/check/cpp/operators.h"
 #include "toolchain/check/cpp/type_mapping.h"
 #include "toolchain/diagnostics/diagnostic_emitter.h"
 #include "toolchain/sem_ir/function.h"
@@ -80,18 +81,10 @@ static auto AddOverloadCandidataes(clang::Sema& sema,
   }
 }
 
-// Returns whether the decl is an operator member function.
-static auto IsOperatorMethodDecl(clang::Decl* decl) -> bool {
-  auto* clang_method_decl = dyn_cast<clang::CXXMethodDecl>(decl);
-  return clang_method_decl && clang_method_decl->isOverloadedOperator();
-}
-
-// Resolve which function to call, or returns an error instruction if overload
-// resolution failed.
-static auto ResolveCalleeId(Context& context, SemIR::LocId loc_id,
-                            SemIR::CppOverloadSetId overload_set_id,
-                            SemIR::InstId self_id,
-                            llvm::ArrayRef<SemIR::InstId> arg_ids)
+auto PerformCppOverloadResolution(Context& context, SemIR::LocId loc_id,
+                                  SemIR::CppOverloadSetId overload_set_id,
+                                  SemIR::InstId self_id,
+                                  llvm::ArrayRef<SemIR::InstId> arg_ids)
     -> SemIR::InstId {
   // Register an annotation scope to flush any Clang diagnostics when we return.
   // This is important to ensure that Clang diagnostics are properly interleaved
@@ -141,7 +134,7 @@ static auto ResolveCalleeId(Context& context, SemIR::LocId loc_id,
           context, loc_id, best_viable_fn->Function,
           // If this is an operator method, the first arg will be used as self.
           arg_exprs.size() -
-              (IsOperatorMethodDecl(best_viable_fn->Function) ? 1 : 0));
+              (IsCppOperatorMethodDecl(best_viable_fn->Function) ? 1 : 0));
       return result;
     }
     case clang::OverloadingResult::OR_No_Viable_Function: {
@@ -166,48 +159,6 @@ static auto ResolveCalleeId(Context& context, SemIR::LocId loc_id,
           GetCppName(context, overload_set.name_id), candidate_set,
           best_viable_fn->Function, arg_exprs);
       return SemIR::ErrorInst::InstId;
-    }
-  }
-}
-
-// Returns whether the function is an imported C++ operator member function.
-static auto IsCppOperatorMethod(Context& context, SemIR::FunctionId function_id)
-    -> bool {
-  SemIR::ClangDeclId clang_decl_id =
-      context.functions().Get(function_id).clang_decl_id;
-  return clang_decl_id.has_value() &&
-         IsOperatorMethodDecl(
-             context.clang_decls().Get(clang_decl_id).key.decl);
-}
-
-auto PerformCallToCppOverloadFunction(Context& context, SemIR::LocId loc_id,
-                                      SemIR::CppOverloadSetId overload_set_id,
-                                      SemIR::InstId self_id,
-                                      llvm::ArrayRef<SemIR::InstId> arg_ids)
-    -> SemIR::InstId {
-  SemIR::InstId callee_id =
-      ResolveCalleeId(context, loc_id, overload_set_id, self_id, arg_ids);
-  SemIR::Callee callee = GetCallee(context.sem_ir(), callee_id);
-  CARBON_KIND_SWITCH(callee) {
-    case CARBON_KIND(SemIR::CalleeError _): {
-      return SemIR::ErrorInst::InstId;
-    }
-    case CARBON_KIND(SemIR::CalleeFunction fn): {
-      CARBON_CHECK(!fn.self_id.has_value());
-      if (self_id.has_value()) {
-        // Preserve the `self` argument from the original callee.
-        fn.self_id = self_id;
-      } else if (IsCppOperatorMethod(context, fn.function_id)) {
-        // Adjust `self` and args for C++ overloaded operator methods.
-        fn.self_id = arg_ids.consume_front();
-      }
-      return PerformCallToFunction(context, loc_id, callee_id, fn, arg_ids);
-    }
-    case CARBON_KIND(SemIR::CalleeCppOverloadSet _): {
-      CARBON_FATAL("overloads can't be recursive");
-    }
-    case CARBON_KIND(SemIR::CalleeNonFunction _): {
-      CARBON_FATAL("overloads should produce functions");
     }
   }
 }
